requirements.txt
streamlit==1.32.2
pypdf==4.1.0
langchain==0.1.13
python-dotenv==1.0.1
chromadb==0.5.0
langchain-google-genai==0.0.11
sentence-transformers>=2.2.2
torch==2.5.1
transformers>=4.30.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13